make table gathering reusable
